split viewsets from views
